@@ -22,6 +22,7 @@ using LibreHardwareMonitor.Hardware.Memory;
 using LibreHardwareMonitor.Hardware.Motherboard;
 using LibreHardwareMonitor.Hardware.Network;
 using LibreHardwareMonitor.Hardware.Psu.Corsair;
+using LibreHardwareMonitor.Hardware.Psu.Msi;
 using LibreHardwareMonitor.Hardware.Storage;
 
 namespace LibreHardwareMonitor.Hardware;
@@ -34,7 +35,7 @@ public class Computer : IComputer
     private readonly List<IGroup> _groups = new();
     private readonly object _lock = new();
     private readonly ISettings _settings;
-        
+
     private bool _batteryEnabled;
     private bool _controllerEnabled;
     private bool _cpuEnabled;
@@ -46,7 +47,6 @@ public class Computer : IComputer
     private bool _psuEnabled;
     private SMBios _smbios;
     private bool _storageEnabled;
-    private bool _ring0Enabled = true;
 
     /// <summary>
     /// Creates a new <see cref="IComputer" /> instance with basic initial <see cref="Settings" />.
@@ -172,7 +172,9 @@ public class Computer : IComputer
                 {
                     Add(new AmdGpuGroup(_settings));
                     Add(new NvidiaGroup(_settings));
-                    Add(new IntelGpuGroup(GetIntelCpus(), _settings));
+
+                    if (_cpuEnabled)
+                        Add(new IntelGpuGroup(GetIntelCpus(), _settings));
                 }
                 else
                 {
@@ -251,10 +253,12 @@ public class Computer : IComputer
                 if (value)
                 {
                     Add(new CorsairPsuGroup(_settings));
+                    Add(new MsiPsuGroup(_settings));
                 }
                 else
                 {
                     RemoveType<CorsairPsuGroup>();
+                    RemoveType<MsiPsuGroup>();
                 }
             }
 
@@ -277,16 +281,6 @@ public class Computer : IComputer
             }
 
             _storageEnabled = value;
-        }
-    }
-
-    /// <inheritdoc />
-    public bool IsRing0Enabled
-    {
-        get { return _ring0Enabled; }
-        set
-        {
-            _ring0Enabled = value;
         }
     }
 
@@ -330,14 +324,6 @@ public class Computer : IComputer
             w.Write("Process Type: ");
             w.WriteLine(IntPtr.Size == 4 ? "32-Bit" : "64-Bit");
             w.WriteLine();
-
-            string r = Ring0.GetReport();
-            if (r != null)
-            {
-                NewSection(w);
-                w.Write(r);
-                w.WriteLine();
-            }
 
             NewSection(w);
             w.WriteLine("Sensors");
@@ -474,7 +460,7 @@ public class Computer : IComputer
 
     private void RemoveType<T>() where T : IGroup
     {
-        List<T> list = new();
+        List<T> list = [];
 
         lock (_lock)
         {
@@ -490,7 +476,7 @@ public class Computer : IComputer
     }
 
     /// <summary>
-    /// If hasn't been opened before, opens <see cref="SMBios" />, <see cref="Ring0" />, <see cref="OpCode" /> and triggers the private <see cref="AddGroups" /> method depending on which categories are
+    /// If hasn't been opened before, opens <see cref="SMBios" />, <see cref="OpCode" /> and triggers the private <see cref="AddGroups" /> method depending on which categories are
     /// enabled.
     /// </summary>
     public void Open()
@@ -499,9 +485,6 @@ public class Computer : IComputer
             return;
 
         _smbios = new SMBios();
-
-        if (IsRing0Enabled)
-            Ring0.Open();
 
         Mutexes.Open();
         OpCode.Open();
@@ -526,7 +509,9 @@ public class Computer : IComputer
         {
             Add(new AmdGpuGroup(_settings));
             Add(new NvidiaGroup(_settings));
-            Add(new IntelGpuGroup(GetIntelCpus(), _settings));
+
+            if (_cpuEnabled)
+                Add(new IntelGpuGroup(GetIntelCpus(), _settings));
         }
 
         if (_controllerEnabled)
@@ -546,7 +531,10 @@ public class Computer : IComputer
             Add(new NetworkGroup(_settings));
 
         if (_psuEnabled)
+        {
             Add(new CorsairPsuGroup(_settings));
+            Add(new MsiPsuGroup(_settings));
+        }
 
         if (_batteryEnabled)
             Add(new BatteryGroup(_settings));
@@ -627,7 +615,7 @@ public class Computer : IComputer
     }
 
     /// <summary>
-    /// If opened before, removes all <see cref="IGroup" /> and triggers <see cref="OpCode.Close" />, <see cref="InpOut.Close" /> and <see cref="Ring0.Close" />.
+    /// If opened before, removes all <see cref="IGroup" /> and triggers <see cref="OpCode.Close" />.
     /// </summary>
     public void Close()
     {
@@ -644,8 +632,6 @@ public class Computer : IComputer
         }
 
         OpCode.Close();
-        InpOut.Close();
-        Ring0.Close();
         Mutexes.Close();
 
         _smbios = null;
