@@ -196,11 +196,13 @@ public class Motherboard : IHardware
     /// </summary>
     public void Close()
     {
-        _lmSensors?.Close();
         foreach (IHardware iHardware in SubHardware)
         {
             if (iHardware is Hardware hardware)
                 hardware.Close();
         }
+
+        _lmSensors?.Close();
+        _lpcIO?.Close();
     }
 }
