@@ -4,7 +4,7 @@
 using System;
 using System.Collections.Generic;
 using System.Text;
-using System.Threading;
+using LibreHardwareMonitor.PawnIo;
 
 // ReSharper disable InconsistentNaming
 
@@ -12,13 +12,9 @@ namespace LibreHardwareMonitor.Hardware;
 
 internal class RyzenSMU
 {
-    private const byte SMU_PCI_ADDR_REG = 0xC4;
-    private const byte SMU_PCI_DATA_REG = 0xC8;
-    private const uint SMU_REQ_MAX_ARGS = 6;
-    private const uint SMU_RETRIES_MAX = 8096;
-
     private readonly CpuCodeName _cpuCodeName;
     private readonly bool _supportedCPU;
+    private readonly Exception _unsupportedCPUException;
 
     private readonly Dictionary<uint, Dictionary<uint, SmuSensorType>> _supportedPmTableVersions = new()
     {
@@ -148,59 +144,35 @@ internal class RyzenSMU
         }
     };
 
-    private uint _argsAddr;
-    private uint _cmdAddr;
-    private uint _dramAddrHi;
-    private uint _dramBaseAddr;
     private uint _pmTableSize;
     private uint _pmTableSizeAlt;
     private uint _pmTableVersion;
-    private uint _rspAddr;
+    private uint _dramBaseAddr;
 
-    public RyzenSMU(uint family, uint model, uint packageType)
+    private readonly RyzenSmu _ryzenSmu;
+
+    public RyzenSMU()
     {
-        _cpuCodeName = GetCpuCodeName(family, model, packageType);
-
-        _supportedCPU = Environment.Is64BitOperatingSystem == Environment.Is64BitProcess && SetAddresses(_cpuCodeName);
-
-        if (_supportedCPU && InpOut.Open())
-            SetupPmTableAddrAndSize();
-    }
-
-    private static CpuCodeName GetCpuCodeName(uint family, uint model, uint packageType)
-    {
-        return family switch
+        try
         {
-            0x17 => model switch
-            {
-                0x01 => packageType == 7 ? CpuCodeName.Threadripper : CpuCodeName.SummitRidge,
-                0x08 => packageType == 7 ? CpuCodeName.Colfax : CpuCodeName.PinnacleRidge,
-                0x11 => CpuCodeName.RavenRidge,
-                0x18 => packageType == 2 ? CpuCodeName.RavenRidge2 : CpuCodeName.Picasso,
-                0x20 => CpuCodeName.Dali,
-                0x31 => CpuCodeName.CastlePeak,
-                0x60 => CpuCodeName.Renoir,
-                0x71 => CpuCodeName.Matisse,
-                0x90 => CpuCodeName.Vangogh,
-                _ => CpuCodeName.Undefined
-            },
-            0x19 => model switch
-            {
-                0x00 => CpuCodeName.Milan,
-                0x20 or 0x21 => CpuCodeName.Vermeer,
-                0x40 => CpuCodeName.Rembrandt,
-                0x50 => CpuCodeName.Cezanne,
-                0x61 => CpuCodeName.Raphael,
-                _ => CpuCodeName.Undefined
-            },
-            0x1A => model switch
-            {
-                0x44 => CpuCodeName.GraniteRidge,
-                _ => CpuCodeName.Undefined
-            },
-            _ => CpuCodeName.Undefined
-        };
+            _ryzenSmu = new RyzenSmu();
+
+            _cpuCodeName = (CpuCodeName)_ryzenSmu.GetCodeName();
+
+            _ryzenSmu.ResolvePmTable(out _pmTableVersion, out _dramBaseAddr);
+
+            SetupPmTableSize();
+
+            _supportedCPU = true;
+        }
+        catch (Exception e)
+        {
+            _supportedCPU = false;
+            _unsupportedCPUException = e;
+        }
     }
+
+    public void Close() => _ryzenSmu.Close();
 
     public string GetReport()
     {
@@ -218,62 +190,28 @@ internal class RyzenSMU
             r.AppendLine($" PM table start address: 0x{_dramBaseAddr:X8}");
             r.AppendLine();
             r.AppendLine(" PM table dump:");
-            r.AppendLine("  Idx    Offset   Value");
-            r.AppendLine(" ------------------------");
 
-            float[] pm_values = GetPmTable();
-            for (int i = 0; i < pm_values.Length; i++)
+            try
             {
-                r.AppendLine($" {i,4}    0x{i * 4:X3}    {pm_values[i]}");
+                float[] pm_values = UpdateAndReadDram();
+                r.AppendLine("  Idx    Offset   Value");
+                r.AppendLine(" ------------------------");
+                for (int i = 0; i < pm_values.Length; i++)
+                {
+                    r.AppendLine($" {i,4}    0x{i * 4:X3}    {pm_values[i]}");
+                }
             }
+            catch (Exception e)
+            {
+                r.AppendLine($" Exception: {e.Message}");
+            }
+        }
+        else
+        {
+            r.AppendLine($" Initialization exception: {_unsupportedCPUException.Message}");
         }
 
         return r.ToString();
-    }
-
-    private bool SetAddresses(CpuCodeName codeName)
-    {
-        switch (codeName)
-        {
-            case CpuCodeName.CastlePeak:
-            case CpuCodeName.Matisse:
-            case CpuCodeName.Vermeer:
-            case CpuCodeName.Raphael:
-            case CpuCodeName.GraniteRidge:
-                _cmdAddr = 0x3B10524;
-                _rspAddr = 0x3B10570;
-                _argsAddr = 0x3B10A40;
-                return true;
-
-            case CpuCodeName.Colfax:
-            case CpuCodeName.SummitRidge:
-            case CpuCodeName.Threadripper:
-            case CpuCodeName.PinnacleRidge:
-                _cmdAddr = 0x3B1051C;
-                _rspAddr = 0x3B10568;
-                _argsAddr = 0x3B10590;
-                return true;
-
-            case CpuCodeName.Renoir:
-            case CpuCodeName.Picasso:
-            case CpuCodeName.RavenRidge:
-            case CpuCodeName.RavenRidge2:
-            case CpuCodeName.Dali:
-                _cmdAddr = 0x3B10A20;
-                _rspAddr = 0x3B10A80;
-                _argsAddr = 0x3B10A88;
-                return true;
-
-            default:
-                return false;
-        }
-    }
-
-    public uint GetSmuVersion()
-    {
-        uint[] args = [1];
-
-        return SendCommand(0x02, ref args) ? args[0] : 0;
     }
 
     public Dictionary<uint, SmuSensorType> GetPmTableStructure()
@@ -291,51 +229,44 @@ internal class RyzenSMU
 
     public float[] GetPmTable()
     {
-        if (!_supportedCPU || !TransferTableToDram())
+        if (!_supportedCPU)
             return [0];
 
-        float[] table = ReadDramToArray();
-
-        // Fix for Zen+ empty values on first call.
-        if (table.Length == 0 || table[0] == 0)
+        float[] table = null;
+        for (int tries_left = 2; tries_left != 0; --tries_left)
         {
-            Thread.Sleep(10);
-            TransferTableToDram();
-            table = ReadDramToArray();
+            table = null;
+            try
+            {
+                table = UpdateAndReadDram();
+            }
+            catch
+            {
+                // ignored
+            }
+
+            if (table is { Length: > 0 } && table[0] != 0)
+            {
+                return table;
+            }
         }
 
-        return table;
+        return table is { Length: > 0 } ? table : [0];
     }
 
-    private float[] ReadDramToArray()
+    private float[] UpdateAndReadDram()
     {
         float[] table = new float[_pmTableSize / 4];
 
-        IntPtr pMemory = Environment.Is64BitProcess ? new IntPtr(_dramBaseAddr | (long)_dramAddrHi << 32) : new IntPtr(_dramBaseAddr);
-
-        byte[] bytes = InpOut.ReadMemory(pMemory, _pmTableSize);
-        if (bytes != null)
-            Buffer.BlockCopy(bytes, 0, table, 0, bytes.Length);
+        _ryzenSmu.UpdatePmTable();
+        long[] read = _ryzenSmu.ReadPmTable((int)((_pmTableSize + 7) / 8));
+        Buffer.BlockCopy(read, 0, table, 0, (int)_pmTableSize);
 
         return table;
-    }
-
-    private bool SetupPmTableAddrAndSize()
-    {
-        if (_pmTableSize == 0)
-            SetupPmTableSize();
-
-        if (_dramBaseAddr == 0)
-            SetupDramBaseAddr();
-
-        return _dramBaseAddr != 0 && _pmTableSize != 0;
     }
 
     private void SetupPmTableSize()
     {
-        if (!GetPmTableVersion(ref _pmTableVersion))
-            return;
-
         switch (_cpuCodeName)
         {
             case CpuCodeName.Matisse:
@@ -454,288 +385,6 @@ internal class RyzenSMU
         }
     }
 
-    private bool GetPmTableVersion(ref uint version)
-    {
-        uint[] args = [0];
-        uint fn;
-
-        switch (_cpuCodeName)
-        {
-            case CpuCodeName.RavenRidge:
-            case CpuCodeName.Picasso:
-                fn = 0x0c;
-                break;
-            case CpuCodeName.Matisse:
-            case CpuCodeName.Vermeer:
-                fn = 0x08;
-                break;
-            case CpuCodeName.Renoir:
-                fn = 0x06;
-                break;
-            case CpuCodeName.Raphael:
-            case CpuCodeName.GraniteRidge:
-                fn = 0x05;
-                break;
-            default:
-                return false;
-        }
-
-        bool ret = SendCommand(fn, ref args);
-        version = args[0];
-
-        return ret;
-    }
-
-    private void SetupAddrClass1(uint[] fn)
-    {
-        uint[] args = [1, 1];
-
-        bool command = SendCommand(fn[0], ref args);
-        if (!command)
-            return;
-
-        _dramBaseAddr = args[0];
-        _dramAddrHi = args[1];
-    }
-
-    private void SetupAddrClass2(uint[] fn)
-    {
-        uint[] args = [0, 0, 0, 0, 0, 0];
-
-        bool command = SendCommand(fn[0], ref args);
-        if (!command)
-            return;
-
-        args = [0];
-        command = SendCommand(fn[1], ref args);
-        if (!command)
-            return;
-
-        _dramBaseAddr = args[0];
-    }
-
-    private void SetupAddrClass3(uint[] fn)
-    {
-        uint[] parts = [0, 0];
-
-        // == Part 1 ==
-        uint[] args = [3];
-        bool command = SendCommand(fn[0], ref args);
-        if (!command)
-            return;
-
-        args = [3];
-        command = SendCommand(fn[2], ref args);
-        if (!command)
-            return;
-
-        // 1st Base.
-        parts[0] = args[0];
-        // == Part 1 End ==
-
-        // == Part 2 ==
-        args = [3];
-        command = SendCommand(fn[1], ref args);
-        if (!command)
-            return;
-
-        args = [5];
-        command = SendCommand(fn[0], ref args);
-        if (!command)
-            return;
-
-        args = [5];
-        command = SendCommand(fn[2], ref args);
-        if (!command)
-            return;
-
-        // 2nd base.
-        parts[1] = args[0];
-        // == Part 2 End ==
-
-        _dramBaseAddr = parts[0] & 0xFFFFFFFF;
-    }
-
-    private void SetupDramBaseAddr()
-    {
-        uint[] fn = [0, 0, 0];
-
-        switch (_cpuCodeName)
-        {
-            case CpuCodeName.Raphael:
-            case CpuCodeName.GraniteRidge:
-                fn[0] = 0x04;
-                SetupAddrClass1(fn);
-                return;
-            case CpuCodeName.Vermeer:
-            case CpuCodeName.Matisse:
-            case CpuCodeName.CastlePeak:
-                fn[0] = 0x06;
-                SetupAddrClass1(fn);
-                return;
-            case CpuCodeName.Renoir:
-                fn[0] = 0x66;
-                SetupAddrClass1(fn);
-                return;
-            case CpuCodeName.Colfax:
-            case CpuCodeName.PinnacleRidge:
-                fn[0] = 0x0b;
-                fn[1] = 0x0c;
-                SetupAddrClass2(fn);
-                return;
-            case CpuCodeName.Dali:
-            case CpuCodeName.Picasso:
-            case CpuCodeName.RavenRidge:
-            case CpuCodeName.RavenRidge2:
-                fn[0] = 0x0a;
-                fn[1] = 0x3d;
-                fn[2] = 0x0b;
-                SetupAddrClass3(fn);
-                return;
-            default:
-                return;
-        }
-    }
-
-    public bool TransferTableToDram()
-    {
-        uint[] args = [0];
-        uint fn;
-
-        switch (_cpuCodeName)
-        {
-            case CpuCodeName.Raphael:
-            case CpuCodeName.GraniteRidge:
-                fn = 0x03;
-                break;
-            case CpuCodeName.Matisse:
-            case CpuCodeName.Vermeer:
-                fn = 0x05;
-                break;
-            case CpuCodeName.Renoir:
-                args[0] = 3;
-                fn = 0x65;
-                break;
-            case CpuCodeName.Picasso:
-            case CpuCodeName.RavenRidge:
-            case CpuCodeName.RavenRidge2:
-                args[0] = 3;
-                fn = 0x3d;
-                break;
-            default:
-                return false;
-        }
-
-        return SendCommand(fn, ref args);
-    }
-
-    private bool SendCommand(uint msg, ref uint[] args)
-    {
-        uint[] cmdArgs = new uint[SMU_REQ_MAX_ARGS];
-        int argsLength = Math.Min(args.Length, cmdArgs.Length);
-
-        for (int i = 0; i < argsLength; ++i)
-            cmdArgs[i] = args[i];
-
-        uint tmp = 0;
-        if (Mutexes.WaitPciBus(5000))
-        {
-            // Step 1: Wait until the RSP register is non-zero.
-
-            tmp = 0;
-            uint retries = SMU_RETRIES_MAX;
-            do
-            {
-                if (!ReadReg(_rspAddr, ref tmp))
-                {
-                    Mutexes.ReleasePciBus();
-                    return false;
-                }
-            }
-            while (tmp == 0 && retries-- != 0);
-
-            // Step 1.b: A command is still being processed meaning a new command cannot be issued.
-
-            if (retries == 0 && tmp == 0)
-            {
-                Mutexes.ReleasePciBus();
-                return false;
-            }
-
-            // Step 2: Write zero (0) to the RSP register
-            WriteReg(_rspAddr, 0);
-
-            // Step 3: Write the argument(s) into the argument register(s)
-            for (int i = 0; i < cmdArgs.Length; ++i)
-                WriteReg(_argsAddr + (uint)(i * 4), cmdArgs[i]);
-
-            // Step 4: Write the message Id into the Message ID register
-            WriteReg(_cmdAddr, msg);
-
-            // Step 5: Wait until the Response register is non-zero.
-            tmp = 0;
-            retries = SMU_RETRIES_MAX;
-            do
-            {
-                if (!ReadReg(_rspAddr, ref tmp))
-                {
-                    Mutexes.ReleasePciBus();
-                    return false;
-                }
-            }
-            while (tmp == 0 && retries-- != 0);
-
-            if (retries == 0 && tmp != (uint)Status.OK)
-            {
-                Mutexes.ReleasePciBus();
-                return false;
-            }
-
-            // Step 6: If the Response register contains OK, then SMU has finished processing  the message.
-
-            args = new uint[SMU_REQ_MAX_ARGS];
-            for (byte i = 0; i < SMU_REQ_MAX_ARGS; i++)
-            {
-                if (!ReadReg(_argsAddr + (uint)(i * 4), ref args[i]))
-                {
-                    Mutexes.ReleasePciBus();
-                    return false;
-                }
-            }
-
-            ReadReg(_rspAddr, ref tmp);
-            Mutexes.ReleasePciBus();
-        }
-
-        return tmp == (uint)Status.OK;
-    }
-
-    private static void WriteReg(uint addr, uint data)
-    {
-        if (Mutexes.WaitPciBus(10))
-        {
-            if (Ring0.WritePciConfig(0x00, SMU_PCI_ADDR_REG, addr))
-                Ring0.WritePciConfig(0x00, SMU_PCI_DATA_REG, data);
-
-            Mutexes.ReleasePciBus();
-        }
-    }
-
-    private static bool ReadReg(uint addr, ref uint data)
-    {
-        bool read = false;
-
-        if (Mutexes.WaitPciBus(10))
-        {
-            if (Ring0.WritePciConfig(0x00, SMU_PCI_ADDR_REG, addr))
-                read = Ring0.ReadPciConfig(0x00, SMU_PCI_DATA_REG, out data);
-
-            Mutexes.ReleasePciBus();
-        }
-
-        return read;
-    }
-
     public struct SmuSensorType
     {
         public string Name;
@@ -743,18 +392,10 @@ internal class RyzenSMU
         public float Scale;
     }
 
-    private enum Status : uint
-    {
-        OK = 0x01,
-        CmdRejectedBusy = 0xFC,
-        CmdRejectedPrereq = 0xFD,
-        UnknownCmd = 0xFE,
-        Failed = 0xFF
-    }
 
     private enum CpuCodeName
     {
-        Undefined,
+        Undefined = -1,
         Colfax,
         Renoir,
         Picasso,
