@@ -4,6 +4,7 @@
 // All Rights Reserved.
 
 using System.Runtime.InteropServices;
+using Windows.Win32.Storage.Nvme;
 using LibreHardwareMonitor.Interop;
 
 namespace LibreHardwareMonitor.Hardware.Storage;
@@ -12,7 +13,7 @@ internal interface INVMeDrive
 {
     SafeHandle Identify(StorageInfo storageInfo);
 
-    bool IdentifyController(SafeHandle hDevice, out Kernel32.NVME_IDENTIFY_CONTROLLER_DATA data);
+    bool IdentifyController(SafeHandle hDevice, out NVME_IDENTIFY_CONTROLLER_DATA data);
 
-    bool HealthInfoLog(SafeHandle hDevice, out Kernel32.NVME_HEALTH_INFO_LOG data);
+    bool HealthInfoLog(SafeHandle hDevice, out NVME_HEALTH_INFO_LOG data);
 }
